requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
blinker==1.4
Brotli==1.0.9
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
coverage==6.3.2
cryptography==37.0.2
decorator==5.1.1
execnet==1.9.0
Faker==13.7.0
flake8==4.0.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.4
hyperframe==6.0.1
idna==2.10
iniconfig==1.1.1
isort==5.10.1
kaitaistruct==0.9
mccabe==0.6.1
oauthlib==3.2.0
outcome==1.2.0
packaging==20.9
pluggy==0.13.1
protobuf==4.21.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydivert==2.1.0
pyflakes==2.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-cov==2.12.1
pytest-custom-exit-code==0.3.0
pytest-forked==1.4.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.3.0
self==2020.12.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.5
webdriver-manager==3.5.4
Werkzeug==2.0.3
wsproto==1.1.0
zstandard==0.17.0
